# docs/api_doc/swagger.json
{"openapi":"3.0.1","info":{"title":"'@project.name@'","description":"'@project.description@'","contact":{"name":"John Doe","email":"dumy@dumy"},"version":"'@project.version@'"},"servers":[{"url":"http://localhost:8080","description":"'@project.description@'"}],"paths":{"/visitor":{"get":{"tags":["visitor-controller"],"operationId":"getAllVisitors","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/VisitorDTO"}}}}}}},"put":{"tags":["visitor-controller"],"operationId":"updateAuthorized","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VisitorRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VisitorDTO"}}}}}}},"/authorized-ranges/register":{"post":{"tags":["authorized-ranges-controller"],"operationId":"registerAuthorizedRange","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterAuthorizationRangesDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthorizedRanges"}}}}}}},"/worker-access":{"get":{"tags":["worker-access-controller"],"operationId":"getWorkerAccess","parameters":[{"name":"worker_id","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"plot_id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"date_from","in":"query","required":true,"schema":{"type":"string","format":"date"}},{"name":"date_to","in":"query","required":true,"schema":{"type":"string","format":"date"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WorkerAccessDTO"}}}}}}}},"/visitor/{docNumber}":{"get":{"tags":["visitor-controller"],"operationId":"getVisitorByDocNumber","parameters":[{"name":"docNumber","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VisitorDTO"}}}}}}},"/visit-access":{"get":{"tags":["visit-access-controller"],"operationId":"getVisitAccesses","parameters":[{"name":"plot","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"visitor_id","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"date_from","in":"query","required":true,"schema":{"type":"string","format":"date"}},{"name":"date_to","in":"query","required":true,"schema":{"type":"string","format":"date"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/VisitAccessDTO"}}}}}}}},"/supplier-access":{"get":{"tags":["supplier-access-controller"],"operationId":"getSupplierAccesses","parameters":[{"name":"supplier_id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"date_from","in":"query","required":true,"schema":{"type":"string","format":"date"}},{"name":"date_to","in":"query","required":true,"schema":{"type":"string","format":"date"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WorkerAccessDTO"}}}}}}}},"/qr/{docNumber}":{"get":{"tags":["qr-controller"],"operationId":"generateQr","parameters":[{"name":"docNumber","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string","format":"byte"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/owner-access":{"get":{"tags":["owner-access-controller"],"operationId":"getOwnerAccesses","parameters":[{"name":"plot_id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"date_from","in":"query","required":true,"schema":{"type":"string","format":"date"}},{"name":"date_to","in":"query","required":true,"schema":{"type":"string","format":"date"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OwnerAccessDTO"}}}}}}}},"/employee-access":{"get":{"tags":["employee-access-controller"],"operationId":"getEmployeeAccesses","parameters":[{"name":"employee_id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"date_from","in":"query","required":true,"schema":{"type":"string","format":"date"}},{"name":"date_to","in":"query","required":true,"schema":{"type":"string","format":"date"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EmployeeAccessDTO"}}}}}}}}},"components":{"schemas":{"VisitorRequestDto":{"type":"object","properties":{"name":{"type":"string"},"last_name":{"type":"string"},"doc_number":{"type":"integer","format":"int64"},"birth_date":{"pattern":"dd-MM-yyyy","type":"string","example":"18-09-2024"},"owner_id":{"type":"integer","format":"int64"},"is_active":{"type":"boolean"}}},"VisitorDTO":{"type":"object","properties":{"visitor_id":{"type":"integer","format":"int64"},"owner_id":{"type":"integer","format":"int64"},"name":{"type":"string"},"last_name":{"type":"string"},"doc_number":{"type":"integer","format":"int64"},"birth_date":{"type":"string","format":"date"},"is_active":{"type":"boolean"}}},"LocalTime":{"type":"object","properties":{"hour":{"type":"integer","format":"int32"},"minute":{"type":"integer","format":"int32"},"second":{"type":"integer","format":"int32"},"nano":{"type":"integer","format":"int32"}}},"RegisterAuthorizationRangesDTO":{"type":"object","properties":{"comment":{"type":"string"},"auth_type_id":{"type":"integer","format":"int64"},"visitor_id":{"type":"integer","format":"int64"},"external_id":{"type":"integer","format":"int64"},"date_from":{"type":"string","format":"date"},"date_to":{"type":"string","format":"date"},"hour_from":{"$ref":"#/components/schemas/LocalTime"},"hour_to":{"$ref":"#/components/schemas/LocalTime"},"day_of_weeks":{"type":"array","items":{"type":"string","enum":["MONDAY","TUESDAY","WEDNESDAY","THURSDAY","FRIDAY","SATURDAY","SUNDAY"]}},"plot_id":{"type":"integer","format":"int64"}}},"AuthorizedRanges":{"type":"object","properties":{"createdUser":{"type":"integer","format":"int64"},"createdDate":{"type":"string","format":"date-time"},"lastUpdatedUser":{"type":"integer","format":"int64"},"lastUpdatedDate":{"type":"string","format":"date-time"},"authRangeId":{"type":"integer","format":"int64"},"authType":{"type":"integer","format":"int64"},"visitorId":{"type":"integer","format":"int64"},"externalId":{"type":"integer","format":"int64"},"dateFrom":{"type":"string","format":"date"},"dateTo":{"type":"string","format":"date"},"hourFrom":{"$ref":"#/components/schemas/LocalTime"},"hourTo":{"$ref":"#/components/schemas/LocalTime"},"days":{"type":"string"},"plotId":{"type":"integer","format":"int64"},"comment":{"type":"string"},"active":{"type":"boolean"}}},"WorkerAccessDTO":{"type":"object","properties":{"external_id":{"type":"integer","format":"int64"},"entry_date_time":{"type":"string","format":"date-time"},"exit_date_time":{"type":"string","format":"date-time"},"car_description":{"type":"string"},"vehicle_reg":{"type":"string"},"vehicle_description":{"type":"string"},"comments":{"type":"string"}}},"VisitAccessDTO":{"type":"object","properties":{"first_name":{"type":"string"},"last_name":{"type":"string"},"doc_number":{"type":"integer","format":"int64"},"entry_date_time":{"type":"string","format":"date-time"},"exit_date_time":{"type":"string","format":"date-time"},"car_description":{"type":"string"},"vehicle_reg":{"type":"string"},"vehicle_description":{"type":"string"},"comments":{"type":"string"}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}},"OwnerAccessDTO":{"type":"object","properties":{"first_name":{"type":"string"},"last_name":{"type":"string"},"entry_date_time":{"type":"string","format":"date-time"},"exit_date_time":{"type":"string","format":"date-time"},"comments":{"type":"string"}}},"EmployeeAccessDTO":{"type":"object","properties":{"entry_date_time":{"type":"string","format":"date-time"},"exit_date_time":{"type":"string","format":"date-time"},"comments":{"type":"string"}}}}}}